not_permissed/mkIndices/initial-es-requirements.txt
mysql-connector-python==8.0.33
elasticsearch==8.2.0
